sbi_rt
Function console_write_byte

pub fn console_write_byte(byte: u8) -> SbiRet

Write a single byte to the debug console.

Blocking function

This is a blocking SBI call, and it will only return after writing the specified byte to the debug console. It will also return with SbiRet::failed() if there are I/O errors.

Return value

The SbiRet.value is set to zero, and the possible return error codes returned in SbiRet.error are shown in the table below:

Return code	Description

SbiRet::success()	Byte written successfully.

SbiRet::failed()	Failed to write the byte due to I/O errors.

This function is defined in RISC-V SBI Specification chapter 12.3.