sbi_rt
Function remote_sfence_vma

pub fn remote_sfence_vma(
    hart_mask: HartMask,
    start_addr: usize,
    size: usize,
) -> SbiRet

Execute SFENCE.VMA instructions for all address spaces on remote harts.

This function instructs the remote harts to execute one or more SFENCE.VMA instructions, covering the range of virtual addresses between start_addr and size.

Return value

The possible return error codes returned in SbiRet.error are shown in the table below:

Return code	Description

SbiRet::success()	A remote fence was sent to all the targeted harts successfully.

SbiRet::invalid_address()	start_addr or size is not valid.

This function is defined in RISC-V SBI Specification chapter 8.2.