Crate sbi_rt

Simple RISC-V SBI runtime primitives

Modules

legacy

Chapter 5. Legacy Extensions (EIDs #0x00 - #0x0F)

Structs

Base

RISC-V SBI Base extension

ColdReboot

Cold reboot as reset type

ExtensionInfo

Information about an SBI extension

Fence

Remote Fence extension

Hsm

Hart State Monitor extension

Ipi

Inter-processor Interrupt extension

NoReason

No reason as reset reason

NonRetentive

Default non-retentive hart suspension

Pmu

Performance Monitoring Unit extension

Reset

System Reset extension

Retentive

Default retentive hart suspension

SbiRet

SBI functions return type

Shutdown

Shutdown as reset type

SystemFailure

System failure as reset reason

Timer

Timer programmer extension

Version

SBI specification version

WarmReboot

Warm reboot as reset type

Traits

ConfigFlags

Flags to configuate performance counter

Extension

An SBI extension

ResetReason

A valid reason for system reset

ResetType

A valid type for system reset

StartFlags

Flags to start performance counter

StopFlags

Flags to stop performance counter

SuspendType

A valid suspend type for hart state monitor

Functions

get_marchid

Return value of marchid register in current environment.

get_mimpid

Return value of mimpid register in current environment.

get_mvendorid

Return value of mvendorid register in current environment.

get_sbi_impl_id

Return the current SBI implementation ID.

get_sbi_impl_version

Return the current SBI implementation version.

get_spec_version

Return the current SBI specification version.

hart_get_status

Get the current status (or HSM state id) of the given hart.

hart_start

Start executing the given hart at specified address in supervisor-mode.

hart_stop

Stop executing the calling hart in supervisor-mode.

hart_suspend

Put the calling hart into suspend or platform specific lower power states.

pmu_counter_config_matching

Find and configure a counter from a set of counters.

pmu_counter_fw_read

Provide the current value of a firmware counter.

pmu_counter_get_info

Get details about the specified counter.

pmu_counter_start

Start or enable a set of counters on the calling hart with the specified initial value.

pmu_counter_stop

Stop or disable a set of counters on the calling hart.

pmu_num_counters

Returns the number of counters, both hardware and firmware.

probe_extension

Probe information about one SBI extension from current environment.

remote_fence_i

Execute FENCE.I instruction on remote harts.

remote_hfence_gvma

Execute HFENCE.GVMA instructions for all virtual machines on remote harts.

remote_hfence_gvma_vmid

Execute virtual machine id based HFENCE.GVMA instructions on remote harts.

remote_hfence_vvma

Execute HFENCE.VVMA for all address spaces in current virtual machine on remote harts.

remote_hfence_vvma_asid

Execute address space based HFENCE.VVMA for current virtual machine on remote harts.

remote_sfence_vma

Execute SFENCE.VMA instructions for all address spaces on remote harts.

remote_sfence_vma_asid

Execute address space based SFENCE.VMA instructions on remote harts.

send_ipi

Send an inter-processor interrupt to all harts defined in hart mask.

set_timer

Programs the clock for next event after an absolute time.

system_reset

Reset the system based on provided reset_type and reset_reason.