Function sbi_rt::hart_start

pub fn hart_start(hartid: usize, start_addr: usize, opaque: usize) -> SbiRet

Start executing the given hart at specified address in supervisor-mode.

This function requests the SBI implementation to start executing the given hart at specified address in supervisor-mode.

This call is asynchronous - more specifically, the sbi_hart_start() may return before target hart starts executing as long as the SBI implemenation is capable of ensuring the return code is accurate.

It is recommended that if the SBI implementation is a platform runtime firmware executing in machine-mode (M-mode) then it MUST configure PMP and other the M-mode state before executing in supervisor-mode.

Parameters

The hartid parameter specifies the target hart which is to be started.

The start_addr parameter points to a runtime-specified physical address, where the hart can start executing in supervisor-mode.

The opaque parameter is a usize value which will be set in the a1 register when the hart starts executing at start_addr.

Behavior

The target hart jumps to supervisor mode at address specified by start_addr with following values in specific registers.

Register Name	Register Value

satp	0

sstatus.SIE	0

a0	hartid

a1	opaque parameter

Return value

The possible return error codes returned in SbiRet.error are shown in the table below:

Return code	Description

SBI_SUCCESS	Hart was previously in stopped state. It will start executing from start_addr.

SBI_ERR_INVALID_ADDRESS	start_addr is not valid possibly due to following reasons: 1. It is not a valid physical address. 2. The address is prohibited by PMP to run in supervisor mode.

SBI_ERR_INVALID_PARAM	hartid is not a valid hartid as corresponding hart cannot started in supervisor mode.

SBI_ERR_ALREADY_AVAILABLE	The given hartid is already started.

SBI_ERR_FAILED	The start request failed for unknown reasons.

This function is defined in RISC-V SBI Specification chapter 9.1.